change: initial refactor
