+++ .sequelizerc
@@ -1,8 +1,8 @@
 const path = require('path');
 
 module.exports = {
-  config: path.resolve('./database/config', 'config.js'),
-  'models-path': path.resolve('./database/models'),
-  'seeders-path': path.resolve('./database/seeders'),
-  'migrations-path': path.resolve('./database/migrations')
+  config: path.resolve('./source/database/config', 'config.js'),
+  'models-path': path.resolve('./source/database/models'),
+  'seeders-path': path.resolve('./source/database/seeders'),
+  'migrations-path': path.resolve('./source/database/migrations')
 };
