+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "start": "node --env-file=.env --watch ./bin/www"
+    "dev": "node --env-file=.env --watch ./source/bin/www"
   },
   "dependencies": {
     "bcrypt": "^5.1.1",
@@ -20,5 +20,8 @@
     "mysql2": "^3.13.0",
     "sequelize": "^6.37.6",
     "uuid": "^11.0.5"
+  },
+  "devDependencies": {
+    "sequelize-cli": "^6.6.2"
   }
 }
