+++ source/app.js
@@ -32,7 +32,7 @@ app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser());
 app.use(express.static(path.join(__dirname, '..', 'public')));
-app.use('/images', express.static(path.join(__dirname, 'public/images'))); // Servir el directorio 'public/images'
+app.use('/images', express.static(path.join(__dirname, '../public/images'))); // Servir el directorio 'public/images'
 
 app.use(methodOverride('_method'));
