+++ source/bin/www
@@ -6,7 +6,7 @@
 
 require('dotenv').config();
 
-const app = require('../source/app')
+const app = require('../app')
 var debug = require('debug')('desafioprofesional:server');
 var http = require('http');
