+++ source/database/config/config.js
@@ -6,6 +6,7 @@ module.exports = {
     "password": process.env.DB_PASSWORD,
     "database": process.env.DB_NAME,
     "host": process.env.DB_HOST,
+    "port" : process.env.DB_PORT,
     "dialect": "mysql"
   },
   "test": {
