+++ source/database/models/ProductIngredient.js
